+++ src/mode_testing_holdout.py
@@ -113,9 +113,6 @@ if __name__ == "__main__":
         scorer_dict = {k: scorer_dict[k] for k in config_dict["ml"]["scorer_list"]}
         omicLogger.info("Scorers extracted. Defining plots...")
 
-        # See what plots are defined
-        plot_dict = mode_plotting.define_plots(config_dict["ml"]["problem_type"])
-
         # Pickling doesn't inherit the self.__class__.__dict__, just self.__dict__
         # So set that up here
         # Other option is to modify cls.__getstate__
@@ -192,7 +189,6 @@ if __name__ == "__main__":
             config_dict,
             experiment_folder,
             features_names,
-            plot_dict,
             x,
             y,
             x_train,